+++ comm/p2p/libp2p_test.go
@@ -13,7 +13,7 @@ import (
 	mock_host "github.com/ChainSafe/sygma-relayer/comm/p2p/mock/host"
 	mock_network "github.com/ChainSafe/sygma-relayer/comm/p2p/mock/stream"
 	"github.com/ChainSafe/sygma-relayer/topology"
-	"github.com/ChainSafe/sygma-relayer/tss/common"
+	"github.com/ChainSafe/sygma-relayer/tss/ecdsa/common"
 	"github.com/golang/mock/gomock"
 	"github.com/libp2p/go-libp2p/core/crypto"
 	"github.com/libp2p/go-libp2p/core/host"
