+++ tss/ecdsa/common/base_test.go
@@ -13,8 +13,8 @@ import (
 
 	"github.com/ChainSafe/sygma-relayer/comm"
 	mock_communication "github.com/ChainSafe/sygma-relayer/comm/mock"
-	"github.com/ChainSafe/sygma-relayer/tss/common"
-	mock_tss "github.com/ChainSafe/sygma-relayer/tss/common/mock"
+	"github.com/ChainSafe/sygma-relayer/tss/ecdsa/common"
+	mock_tss "github.com/ChainSafe/sygma-relayer/tss/ecdsa/common/mock"
 	"github.com/binance-chain/tss-lib/tss"
 	"github.com/golang/mock/gomock"
 	"github.com/libp2p/go-libp2p/core/peer"
